requirements.txt
spotipy==2.23.0
flask==3.0.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
python-dotenv>=0.19.0
requests>=2.25.0
joblib>=1.1.0
